WIP: Major database rework
I have switched us from django to python-mongoengine.  This switch
was not made lightly, as I will need to rewrite almost 500 lines of
code.  Still, I think it is for the best.  The current state of django
nonrel (the non relational django fork) is not really ready for python
3.  We NEED python 3 to get decent support for websockets, which are
NEEDED for decent latency between client and server.  Thus, the main
reasons for the switch:

* Django does not like mongo + python 3. python-mongoengine does, tho
with some difficulties

* Django does not really offer anything to this project which
mongoengine doesn't (except a large plugin library).

* mongoengine offers document style storage instead of relational, which
is much better suited to storing the kind of data we are working with here.

* mongo *should* be faster than MySQL or even sqlite for these jobs.

* mongoengine supports easy conversion from bson to json and back, which
is very nice for the front end.